docker: Add build deps for perf
The perf tool is included in the kernel source and has some entertainingly
extensive build dependency options. In preparation for building perf and
it's testsuite add these to our images, more work will be needed to enable
both builds but this will be a bit easier if the dependencies are in the
docker images.

There are some duplicate dependencies with kselftest, I've left these in
place as apt will cope with the duplicate adds without trouble and there
is a documenting effect from including them twice.

Signed-off-by: Mark Brown <broonie@kernel.org>